LayoutTests/fast/scrolling/mac/scrollend-event-on-select-element.html

<!DOCTYPE html> <!-- webkit-test-runner [ AsyncOverflowScrollingEnabled=true ] -->
<html>
<head>
    <style>
        body {
            height: 2000px;
        }
        .scroller {
            position: absolute;
            top: 310px;
            left: 10px;
            height: 300px;
            width: 300px;
            border: 20px solid gray;
            padding: 5px;
        }
        .content {
            width: 200%;
            height: 300%;
        }
        
    </style>
    <script src="../../../resources/js-test.js"></script>
    <script src="../../../resources/ui-helper.js"></script>
    <script>
        var jsTestIsAsync = true;

        var scroller;
        var overflowScrollendEventCount = 0;
        var windowScrollendEventCount = 0;

        async function resetScrollPositions()
        {
            window.scrollTo(0, 300);
            scroller.scrollTop = 0;
            
            // Wait for scroll events to fire.
            await UIHelper.ensurePresentationUpdate();
        }
        
        async function testScrollOverContent()
        {
            debug('');
            debug('Test scroll over content');
            await resetScrollPositions();
            overflowScrollendEventCount = 0;
            windowScrollendEventCount = 0;

            const wheelEventSquence = {
                "events" : [
                    {
                        type : "wheel",
                        viewX : 100,
                        viewY : 100,
                        deltaY : -10,
                        phase : "began"
                    },
                    {
                        type : "wheel",
                        deltaY : -100,
                        phase : "changed"
                    },
                    {
                        type : "wheel",
                        phase : "ended"
                    },
                    {
                        type : "wheel",
                        deltaY : -100,
                        momentumPhase : "began"
                    },
                    {
                        type : "wheel",
                        deltaY : -80,
                        momentumPhase : "changed"
                    },
                    {
                        type : "wheel",
                        momentumPhase : "ended"
                    }
                ]
            };
            await UIHelper.mouseWheelSequence(wheelEventSquence);
            await UIHelper.renderingUpdate();

            shouldBe('overflowScrollendEventCount == 1', 'true');
            shouldBe('windowScrollendEventCount == 0', 'true');
        }

        async function scrollTest()
        {
            await testScrollOverContent();
            finishJSTest();
        }

        window.addEventListener('load', () => {
            scroller = document.querySelector('.scroller');
            scroller.addEventListener('scrollend', () => {
                ++overflowScrollendEventCount;
            }, false);

            window.addEventListener('scrollend', () => {
                ++windowScrollendEventCount;
            }, false);

            setTimeout(scrollTest, 0);
        }, false);
    </script>
</head>
<body>
    <select id="mySelect" class="scroller" size="5">
        <option value="option1">Option 1</option>
        <option value="option2">Option 2</option>
        <option value="option3">Option 3</option>
        <option value="option4">Option 4</option>
        <option value="option5">Option 5</option>
        <option value="option6">Option 6</option>
        <option value="option7">Option 7</option>
        <option value="option8">Option 8</option>
        <option value="option9">Option 9</option>
        <option value="option10">Option 10</option>
        <option value="option10">Option 10</option>
        <option value="option10">Option 10</option>
        <option value="option10">Option 10</option>
        <option value="option10">Option 10</option>
        <option value="option10">Option 10</option>
        <option value="option10">Option 10</option>
        <option value="option10">Option 10</option>
        <option value="option10">Option 10</option>
        <option value="option10">Option 10</option>
    </select>
    <div class="overlapper"></div>
</body>
</html>
