LayoutTests/fast/scrolling/mac/scrollend-event-on-select-element-expected.txt


Test scroll over content
PASS overflowScrollendEventCount == 1 is true
PASS windowScrollendEventCount == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
